Switch from Vec to SmallVec for holding Phase damage
Results in a ~2x speedup by replacing the small allocs (and reallocs)
with stack memory.

Doing 4M Zulrah goes from ~18.5 to ~8.5 seconds on my machine.

This was the first optimization found by profiling allocations. I had
lazily built out with Vec knowing it would need to be replaced and had
delayed it because the perf had been good enough. Ideally, we would also
remove all of the rotation/phase allocations unless the specific runs
were needed to be stored.